gcc build fix

+++ trunk/src/objects/objectFan.cpp
@@ -27,7 +27,7 @@ void FanObject::Collide(Object* obj) {
 	if (obj->GetProperties().is_player) {
 
 		// if the player is going slowly, slow down the new fan speed
-		float player_factor = std::min(	MAX_PLAYER_SPEED, fabs(obj->GetVelX()) ); 
+		float player_factor = std::min(	MAX_PLAYER_SPEED, (float)fabs(obj->GetVelX()) ); 
 		player_factor /= MAX_PLAYER_SPEED;
 
 		// figure out which way to spin the fan (left/right)
@@ -37,7 +37,7 @@ void FanObject::Collide(Object* obj) {
 
 		// calc the new rotational velocity
 		rotate_velocity = direction * 
-						std::max(DEFAULT_FAN_VELOCITY*player_factor, fabs(rotate_velocity));
+						std::max(DEFAULT_FAN_VELOCITY*player_factor, (float)fabs(rotate_velocity));
 	}
 }
